feat: Reduced quarto image size (2.1GB -> 700MB) (#2)
* Reduced file size by stripping debug symbols & rm docs

* Reduce image size by third (2.1GB -> 700MB)

Achieved using Alpine + multi-stage builds.
Initially I got stuck with the libraries being brought in for `deno`
(a Quarto dependency that ships with it) clashing with the system
libraries and breaking R, but this has been overcome.

The only issue is that Cairo isn't working despite installing `cairo`
and `cairo-dev`, but this shouldn't be an issue as our outputs are HTMLs