Include jgo in cache (#12)
Previously just the maven repository was included in the fake home dir,
so when `scyjava` came to run `ImageJ` it used `jgo` to install further
dependencies that were missing. These dependencies were attempted to be
installed in $HOME, which in Apptainer is either a system folder that is
read only, or the user's mounted $HOME. Rather than keeping a cache in
the user's local $HOME, the `jgo` cache is now bundled with the image,
so it can be stored in a container read-only folder.

Also parameterises dependency versions.